methodToPushDeleteColumnIntoColumnsArrayAddedInConstructorBody

+++ src/components/moviesTable.jsx
@@ -19,19 +19,25 @@ class MoviesTable extends Component {
             content: movie => (
                 <Like liked={movie.liked} onClick={() => this.props.onLike(movie)} />
             )
-        },
-        {
-            key: auth.getCurrentUser() && auth.getCurrentUser().isAdmin ? 'delete' : null,
-            content: auth.getCurrentUser() && auth.getCurrentUser().isAdmin ? movie => (
-                <button
-                    onClick={() => this.props.onDelete(movie)}
-                    className="btn btn-danger btn-sm"
-                >
-                    Delete
-                </button>
-            ) : null
         }
     ];
+
+    constructor() {
+        super();
+        const user = auth.getCurrentUser();
+        if (user && user.isAdmin)
+            this.columns.push({
+                key: 'delete',
+                content: movie => (
+                    <button
+                        onClick={() => this.props.onDelete(movie)}
+                        className="btn btn-danger btn-sm"
+                    >
+                        Delete
+                    </button>
+                )
+            });
+    }
 
     render() {
         const { movies, onSort, sortColumn } = this.props;
